Throw accuracy fix

+++ 3UgersProjekt/src/gameObjects/GameObject.java
@@ -28,7 +28,7 @@ public abstract class GameObject {
 	
 	public int id;
 	static int idCounter = 0;
-	public Vector<Integer> vectorPos= new Vector<Integer>(2); 		// The position for the objects
+	public Vector<Double> vectorPos= new Vector<Double>(2); 		// The position for the objects
 	
 	public int width, height; 				// the width and height of the shapes
 	
@@ -38,9 +38,9 @@ public abstract class GameObject {
 	
 	
 	// Constructor
-	public GameObject(int posX, int posY, int width, int height) {
-		vectorPos.add(posX);
-		vectorPos.add(posY);
+	public GameObject(double posX, double posY, int width, int height) {
+		vectorPos.add((double)posX);
+		vectorPos.add((double)posY);
 		this.width = width;
 		this.height = height;
 		
@@ -74,13 +74,13 @@ public abstract class GameObject {
 	
 	public void collision() {
 		if (vectorPos.get(0)-width/2< 0)
-			vectorPos.set(0,0+width/2);
+			vectorPos.set(0,(double)0+width/2);
 		if (vectorPos.get(0)+width/2>Main.n)
-			vectorPos.set(0, Main.n-width/2);
+			vectorPos.set(0, (double)Main.n-width/2);
 		if (vectorPos.get(1)+height/2 > Main.m)
-			vectorPos.set(1, Main.m-height/2);
+			vectorPos.set(1, (double)Main.m-height/2);
 		if (vectorPos.get(1)-height/2 < 0)
-			vectorPos.set(1, 0+height/2);
+			vectorPos.set(1, (double)0+height/2);
 		
 		
 	}
@@ -91,7 +91,7 @@ public abstract class GameObject {
 	 ***************/
 	
 	// Getters
-	public Vector<Integer> getVectorPos() {
+	public Vector<Double> getVectorPos() {
 		return vectorPos;
 	}
